// src/app/auth/models/provider.model.ts
export interface Provider {
  id: string;
  emailUser: string;
  password: string;
  password2: string;
  nameUser: string;
  lastNameUser: string;
  dniUser: number;
  chargeUser: string;
  phoneUser: number;
  companyName: string;
  companyRuc: number;
  companyAddress: string;
  companyField: string;
  companyLogoURL: string;
  salesRepresentative: string;
  numberOfWorkers: number;
  workersIn?: number;
  manager: string;
  affidavitStatus: boolean;
  affidavitURL: string;
  affidavitValidity: Date;
  covidPlan: boolean;
  covidPlanURL: string;
  covidPlanValidity: Date;
  IPERCStatus: boolean;
  IPERCURL: string;
  IPERCValiity: Date;
  ATSStatus: boolean;
  ATSURL: string;
  ATSValidity: Date;
  LOTOStatus: boolean;
  LOTOURL: string;
  LOTOValiity: Date;
  proceduresStatus: boolean;
  proceduresURL: string;
  proceduresValidity: Date;
  MSDSStatus: boolean;
  MSDSURL: string;
  MSDSValidity: Date;
  createdAt: Date & firebase.default.firestore.Timestamp;
  createdBy: string;
  editedAt: Date;
  editedBy: string;
  status: 'registered' | 'enabled' | 'disabled';
  blockedBy?: string;
  blockDescription?: string;
}